Split off Dynamic patching
- Preemptively split dynamic patching into its own project (if I ever want to use it for another project, it'll be much easier to drop it into a library mod)
-
