--- Common/Languages/English/Keyed/settings.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<LanguageData>
-	<!-- Settings menu -->
-	<DegeneralizeWork.SettingsCategory>De-generalize Work</DegeneralizeWork.SettingsCategory>
-	
-	<DegeneralizeWork.NoOptions>No configuration options currently available.</DegeneralizeWork.NoOptions>
-	
-	<DegeneralizeWork.NeedsRestart>All changes require a restart to take effect</DegeneralizeWork.NeedsRestart>
-
-	<DegeneralizeWork.Primer>Due to the nature of this mod, it may be necessary to modify other mods. Use these settings to alter how those mods are patched, thus favoring compatibility or user experience. Patch modes are as follows:</DegeneralizeWork.Primer>
-
-	<!-- Descriptions -->
-	<DegeneralizeWork.LightTouchDescription>Light touch: the most compatible option, listed mod data is only changed to the minimum extent needed for compatibility.</DegeneralizeWork.LightTouchDescription>
-	<DegeneralizeWork.BestGuessDescription>Best guess: will attempt to seamlessly integrate with the listed mod, but might throw an error if there's a use case it isn't programmed for.</DegeneralizeWork.BestGuessDescription>
-	<DegeneralizeWork.HeavyHandDescription>Heavy hand: orders of magnitude slower than Best guess, but should seamless integrate with the listed mod without throwing errors.</DegeneralizeWork.HeavyHandDescription>
-
-	<!-- Mode names -->
-	<DegeneralizeWork.LightTouch>Light touch</DegeneralizeWork.LightTouch>
-	<DegeneralizeWork.BestGuess>Best guess</DegeneralizeWork.BestGuess>
-	<DegeneralizeWork.HeavyHand>Heavy hand</DegeneralizeWork.HeavyHand>
-
-	<DegeneralizeWork.ModLabel>{0} compatibility mode</DegeneralizeWork.ModLabel>
-	
-	<!-- Reasoning -->
-	<DegeneralizeWork.ModTooltip.vanillaexpanded.skills>VE Skills also adds stats for Tailoring and Sculpting.\nLight touch will to keep the duplicates, while Best guess will attempt to supplant the stats with the ones defined by this mod.</DegeneralizeWork.ModTooltip.vanillaexpanded.skills>
-	
-</LanguageData>
