+++ Source/DegeneralizeWork/DegeneralizeWork.slnx
@@ -28,6 +28,14 @@
       <Property Name="UseLegacyDebugEngines" Value="No" />
     </Properties>
   </Project>
+  <Project Path="../DynamicPatch/AliasDynamicPatch.csproj" Id="c7eda336-a7bf-40f0-a710-2fc6c6131cb4">
+    <BuildType Solution="Release 1.1|*" Project="Release" />
+    <BuildType Solution="Release 1.2|*" Project="Release" />
+    <BuildType Solution="Release 1.3|*" Project="Release" />
+    <BuildType Solution="Release 1.4|*" Project="Release" />
+    <BuildType Solution="Release 1.5|*" Project="Release" />
+    <BuildType Solution="Release Current|*" Project="Release" />
+  </Project>
   <Project Path="DegeneralizeWork.csproj">
     <BuildType Solution="Release|*" Project="Release Current" />
   </Project>
