+++ Source/DynamicPatch/DynamicPatch.cs
@@ -5,11 +5,12 @@ using System.Linq;
 using UnityEngine;
 using Verse;
 
-namespace DegeneralizeWork; // TODO: Rename to DynamicPatch
+namespace AliasDynamicPatch;
 
-public class DWCore : Mod
+public class DynamicPatch(ModContentPack content) : Mod(content)
 {
 	public Settings settings;
+	protected string translatePrefix;
 
 	/// <summary>
 	/// Stores the union of <see cref="PatchOperationModSequence.patchModes"/> from each patch loaded.<br/>
@@ -18,16 +19,10 @@ public class DWCore : Mod
 	/// <remarks>this should every patche, even if it is not applied due to PatchMode criteria</remarks>
 	PackageMode packageMode = new();
 
-	public DWCore(ModContentPack content) : base(content)
-	{
-		settings = GetSettings<Settings>();
-	}
 	public void AddMode(PackageMode patch)
 	{
 		packageMode.Add(patch);
 	}
-
-	public override string SettingsCategory() => "SettingsCategory".PrefixTranslate();
 
 	/// <remarks>
 	/// Dev note: reference <see cref="MainMenuDrawer.DoMainMenuControls(Rect, bool)">DoMainMenuControls</see> (language) for how the base game does float menu buttons
@@ -66,7 +61,7 @@ public class DWCore : Mod
 
 			foreach (var mod in pachableMods)
 			{
-				if (listing.ButtonTextLabeled("ModLabel".PrefixTranslate(mod.mod.Name), settings.GetMode(mod.mod.PackageId).ToString().PrefixTranslate(), tooltip: ("ModTooltip." + mod.mod.PackageId).PrefixTranslate()))
+				if (listing.ButtonTextLabeled("ModLabel".PrefixTranslate(mod.mod.Name), settings.GetMode(mod.mod.PackageId).ToString().PrefixTranslate(), tooltip: $"{translatePrefix}.ModTooltip.{mod.mod.PackageId}".Translate()))
 				{
 					List<FloatMenuOption> options = mod.modes
 						.OrderBy(mode => mode)
