+++ Source/DynamicPatch/Utility.cs
@@ -5,14 +5,14 @@ using System.Text;
 using System.Threading.Tasks;
 using Verse;
 
-namespace DegeneralizeWork;
+namespace AliasDynamicPatch;
 public static class Utility
 {
 	public static int PatchModeCount = AllEnumValues<PatchMode>().Count();
 
 	public static IEnumerable<T> AllEnumValues<T>() where T : Enum => Enum.GetValues(typeof(T)).Cast<T>();
 
-	public static string PrefixTranslate(this string key) => ("DegeneralizeWork." + key).Translate();
+	public static string PrefixTranslate(this string key) => ("AliasDynamicPatch." + key).Translate();
 
-	public static string PrefixTranslate(this string key, NamedArgument arg1) => ("DegeneralizeWork." + key).Translate(arg1);
+	public static string PrefixTranslate(this string key, NamedArgument arg1) => ("AliasDynamicPatch." + key).Translate(arg1);
 }
