+++ VE_Skills/Patches/NonLight_Recipes_Patch.xml
@@ -28,6 +28,7 @@
 					<workSpeedStat>SculptingSpeed</workSpeedStat>
 				</value>
 			</li>
-	
+		</operations>
+	</Operation>
 	<!-- TODO: stat offsets in hedifs are getting missed -->
 </Patch>
